requirements.txt
paramiko>=3.4.0
pyyaml>=6.0
rich>=13.7.0